+++ lib/hyperglot/language.py
@@ -1,8 +1,19 @@
 import logging
-from .parse import parse_chars
-from . import SUPPORTLEVELS
+import unicodedata2
+from .parse import parse_chars, list_unique, parse_marks
+from . import SUPPORTLEVELS, CHARACTER_ATTRIBUTES, MARK_BASE
 
 log = logging.getLogger(__name__)
+
+
+def is_mark(c):
+    if not c:
+        return False
+
+    try:
+        return unicodedata2.category(c).startswith("M")
+    except Exception as e:
+        log.error("Cannot get unicode category of '%s': %s" % (c, str(e)))
 
 
 class Language(dict):
@@ -14,7 +25,7 @@ class Language(dict):
     getters
     """
 
-    def __init__(self, data, iso):
+    def __init__(self, data, iso, parse=True):
         """
         Init a single Language with the data from rosetta.yaml
 
@@ -23,11 +34,44 @@ class Language(dict):
             this a private attribute, not dict items, so it does not get
             printed out when converting this Language back to yaml for output
         """
-        self.update(data)
+
         self.iso = iso
+        self.update(data)
+
+        if parse:
+            self.parse()
 
     def __repr__(self):
         return "Language object '%s'" % self.get_name()
+
+    def parse(self):
+        if "orthographies" in self:
+            for o in self["orthographies"]:
+
+                marks = []
+                if "marks" in o:
+                    marks = parse_chars(
+                        o["marks"], decompose=True, retainDecomposed=True)
+
+                # Parse all character lists
+                for type in CHARACTER_ATTRIBUTES:
+                    if type not in o:
+                        continue
+
+                    # Any marks encountered are moved to 'marks'
+                    # Note that any base+mark precomposed will not return marks
+                    # since we use decompose=False, so only marks that are part
+                    # of unencoded combinations will be returned and added to
+                    # 'marks'
+                    o[type] = parse_chars(
+                        o[type], decompose=False, retainDecomposed=True)
+                    for c in o[type]:
+                        if is_mark(c):
+                            marks.append(c)
+                    o[type] = [c for c in o[type] if not is_mark(c)]
+
+                # No duplicate marks
+                o["marks"] = list_unique(marks)
 
     def get_orthography(self, script=None, status=None):
         """
@@ -75,6 +119,53 @@ class Language(dict):
         # Note for multiple-orthography-primary languages (Serbian, Korean,
         # Japanese) this returns only one orthography!
         return matches[0]
+
+    def get_required_marks(self, ort, level="base"):
+        """
+        Get those marks which are not simply combining marks of the passed in
+        chars, but explicitly listed, meaning they cannot be derived from
+        decomposition
+        """
+
+        marks = parse_marks(ort["marks"]) if "marks" in ort else ""
+
+        chars = []
+        if "base" in ort:
+            chars = chars + ort["base"]
+
+        if level == "base":
+            # When checking the required marks for 'base' remove the
+            # decomposable marks from 'auxiliary' from 'marks'
+            if "auxiliary" in ort:
+                decomposed_aux = parse_marks(ort["auxiliary"])
+                marks = [m for m in marks if m not in decomposed_aux]
+
+        if level == "aux" and "auxiliary" in ort:
+            chars = chars + ort["auxiliary"]
+
+        decomposed = parse_marks(chars)
+        unencoded = parse_marks(chars, decompose=False)
+        marks = [m for m in marks if m not in decomposed]
+
+        return marks + unencoded
+
+    def get_all_marks(self, ort, level="base"):
+        """
+        Get all combining marks from a level, and any explicitly listed marks.
+        For 'base' this needs to subtract implicitly listed marks from only
+        'auxiliary'.
+        """
+        marks = parse_marks(ort["marks"]) if "marks" in ort else []
+        decom_base = parse_marks(ort["base"]) if "base" in ort else []
+        decom_aux = parse_marks(ort["auxiliary"]) if "auxiliary" in ort else []
+
+        if level == "base":
+            only_aux = [m for m in decom_aux if m not in decom_base]
+            marks = [m for m in marks + decom_base if m not in only_aux]
+            return list_unique(marks)
+
+        if level == "aux" and "auxiliary" in ort:
+            return list_unique(marks + decom_base + decom_aux)
 
     def get_name(self, script=None, strict=False):
         if script is not None:
@@ -196,18 +287,25 @@ class Language(dict):
         if combined == []:
             return False
 
+        combined = [c for c in combined if c != MARK_BASE]
+
         return set(parse_chars(combined, decompose=decomposed,
                                retainDecomposed=False))
 
-    def has_support(self, chars, level="base", decomposed=False,
-                    checkAllOrthographies=False,
-                    pruneOrthographies=True):
+    def supported(self,
+                  chars,
+                  level="base",
+                  decomposed=False,
+                  marks=False,
+                  checkAllOrthographies=False,
+                  pruneOrthographies=True):
         """
         Return a dict with language support based on the passed in chars
 
         @param chars set: Set of chars to check against.
         @param level str: Support level for which to check.
         @param decomposed bool: Flag to decompose the passed in chars.
+        @param marks bool: Flag to require all marks.
         @param checkAllOrthographies bool: Flag to check also non-primary
             orthographies from this Language object. False by default.
         @param pruneOthographies bool: Flag to remove non-supported
@@ -243,17 +341,24 @@ class Language(dict):
             base = self.get_orthography_chars(ort, "base",
                                               ignoreMerge=checkAllOrthographies,  # noqa
                                               decomposed=decomposed)
-            # and 'marks'
-            marks = self.get_orthography_chars(ort, "marks",
-                                               ignoreMerge=checkAllOrthographies,  # noqa
-                                               decomposed=decomposed)
 
             if base:
                 if marks:
-                    base = set(list(base) + list(marks))
+                    required_marks_base = self.get_all_marks(ort, "base")
+                else:
+                    required_marks_base = self.get_required_marks(ort, "base")
+
+                if required_marks_base != []:
+                    log.debug("Required base marks for %s: %s" %
+                              (self.iso, required_marks_base))
+                base = set(list(base) + list(required_marks_base))
 
                 script = ort["script"]
                 supported = base.issubset(chars)
+
+                if not supported:
+                    log.debug("Missing from base language %s: %s" %
+                              (self.iso, " ".join(base.difference(chars))))
 
                 if supported:
                     # Only check aux if base is supported to begin with
@@ -264,7 +369,24 @@ class Language(dict):
                                                      ignoreMerge=checkAllOrthographies,  # noqa
                                                      decomposed=decomposed)
                     if level == "aux" and aux:
+                        if marks:
+                            required_marks_aux = self.get_all_marks(
+                                ort, "aux")
+                        else:
+                            required_marks_aux = self.get_required_marks(
+                                ort, "aux")
+
+                        if required_marks_aux != []:
+                            log.debug("Required aux marks for %s: %s" %
+                                      (self.iso, required_marks_aux))
+                        aux = set(list(aux) + list(required_marks_aux))
+
                         supported = aux.issubset(chars)
+
+                        if not supported:
+                            log.debug("Missing aux language %s: %s" %
+                                      (self.iso,
+                                       " ".join(aux.difference(chars))))
 
             if supported:
                 if script not in support: